% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubature-package.R
\docType{package}
\name{cubature-package}
\alias{cubature-package}
\alias{cubature}
\title{Cubature is a package for adaptive multidimensional integration over
hypercubes}
\description{
Cubature is a package for adaptive multidimensional integration over
hypercubes. It is a wrapper around the pure C, GPLed implementation by
Steven G. Johnson available.
}
\author{
C code by Steven G. Johnson, R by Balasubramanian Narasimhan

Maintainer: Balasubramanian Narasimhan\href{mailto:naras@stat.stanford.edu}{naras@stat.stanford.edu}
}
\keyword{package}
