% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubintegrate.R
\name{default_args}
\alias{default_args}
\title{Default arguments for each integration method}
\usage{
default_args()
}
\value{
a named list of parameters for each method.
}
\description{
Since each method has a different set of parameters, this function
returns the default values of all parameters that can be modified
and passed to integration routines.
}
\examples{
default_args()
}
