% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubature-package.R
\docType{package}
\name{cubature-package}
\alias{cubature}
\alias{cubature-package}
\title{Cubature is a package for adaptive multidimensional integration over
hypercubes}
\description{
Cubature is a package for adaptive multidimensional integration over
hypercubes. It is a wrapper around the pure C, GPLed implementation by
Steven G. Johnson available from the URL
\url{http://ab-initio.mit.edu/wiki/index.php/Cubature}.
}
\details{
\tabular{ll}{ Package: \tab cubature\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2009-12-17\cr License: \tab GPL V2 or later\cr LazyLoad:
\tab yes\cr } There is only one function in the package called
\link[cubature]{adaptIntegrate}.
}
\author{
C code by Steven G. Johnson, R by Balasubramanian Narasimhan

Maintainer: Balasubramanian Narasimhan<naras@stat.stanford.edu>
}
\references{
See \url{http://ab-initio.mit.edu/wiki/index.php/Cubature}
}
\keyword{package}

