% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuhre.R
\name{cuhre}
\alias{cuhre}
\title{Experimental interface to newer Cuba library}
\usage{
cuhre(f, lowerLimit, upperLimit, ..., tol = 1e-05, ncomp = 1,
  maxEval = 10^6, absError = 0, key = 0L, nvec = 1L,
  doChecking = FALSE)
}
\arguments{
\item{f}{The function (integrand) to be integrated}

\item{lowerLimit}{The lower limit of integration, a vector for hypercubes}

\item{upperLimit}{The upper limit of integration, a vector for hypercubes}

\item{...}{All other arguments passed to the function f}

\item{tol}{The maximum tolerance, default 1e-5.}

\item{ncomp}{The dimension of the integrand, default 1, bears no relation to
the dimension of the hypercube}

\item{maxEval}{The maximum number of function evaluations needed, default 10^6.
Note that the actual number of function evaluations
performed is only approximately guaranteed not to exceed this number.}

\item{absError}{The maximum absolute error tolerated}

\item{key}{the quadrature rule key, default 0}

\item{nvec}{the number of vectorization points, ignored but fixed at 1 for now}

\item{doChecking}{A flag to be a bit anal about checking inputs to C
routines. A FALSE value results in approximately 9 percent speed gain in our
experiments. Your mileage will of course vary. Default value is FALSE.}
}
\value{
The returned value is a list of five items:
\describe{
  \item{integral}{thevalue of the integral}
   \item{error}{the estimated relative error}
    \item{functionEvaluations}{the number of times the function was evaluated}
    \item{nregions}{the number of regions used}
    \item{returnCode}{the actual integer return code of the C routine}
}
}
\description{
Experimental interface to newer Cuba library
}
\references{
See \url{http://www.feynarts.de/cuba/}
}
\author{
Balasubramanian Narasimhan
}
\keyword{math}
